pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "euclid"
version = "1.0.0"
description = "Euclid current_release runtime and full_vision certification tooling."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=2.3.5,<3",
  "pandas>=2.2.2,<3",
  "Pillow>=11.3.0,<12",
  "pydantic>=2.12.5,<3",
  "pyarrow>=20.0.0,<21",
  "typer>=0.15.2,<1",
  "nbformat>=5.10.4,<6",
  "nbclient>=0.10.2,<1",
  "sqlalchemy>=2.0.38,<2.1",
  "pyinstrument>=5.0.0,<6",
  "PyYAML>=6.0.2,<7",
]

[project.optional-dependencies]
dev = [
  "build==1.2.2.post1",
  "pytest==8.2.0",
  "pytest-cov==6.2.1",
  "ruff==0.6.8",
]

[project.scripts]
euclid = "euclid.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.package-data]
euclid = [
  "_assets/**/*.yaml",
  "_assets/**/*.csv",
  "_assets/**/*.ipynb",
  "_assets/**/*.json",
  "_assets/**/*.html",
  "_assets/**/*.css",
  "_assets/**/*.js",
  "_assets/**/*.mjs",
  "_assets/**/*.woff",
  "_assets/**/*.woff2",
  "_assets/**/*.ttf",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-ra --strict-config --strict-markers --ignore=tests/spec_compiler"
minversion = "8.0"
testpaths = [
  "tests",
]

[tool.ruff]
line-length = 88
target-version = "py311"
src = ["src", "tests", "tools"]

[tool.ruff.lint]
select = ["E", "F", "I"]
